+++ apps/chuni-web/src/routes/(app)/dashboard/+layout.server.ts
@@ -6,7 +6,7 @@ import { db } from "$lib/db";
 import { getUserStats } from "$lib/functions/userStats";
 
 import { isAdminUser } from "@repo/core/web";
-import { apiKey, jobTable, manualRatingTable } from "@repo/database/chuni";
+import { jobTable } from "@repo/database/chuni";
 
 import type { LayoutServerLoad } from "./$types";
 
@@ -17,26 +17,13 @@ export const load: LayoutServerLoad = async ({ parent }) => {
     error(401, "Unauthorized");
   }
 
-  const [jobResult, apiKeyResult, userStats, manualRatings] = await Promise.all(
-    [
-      db
-        .select({ count: count() })
-        .from(jobTable)
-        .where(eq(jobTable.userId, session.user.id)),
-      db
-        .select({ apiKey: apiKey.apiKey, createdAt: apiKey.createdAt })
-        .from(apiKey)
-        .where(eq(apiKey.userId, session.user.id)),
-      getUserStats(session.user.id),
-      db
-        .select({
-          rating: manualRatingTable.rating,
-          timestamp: manualRatingTable.timestamp,
-        })
-        .from(manualRatingTable)
-        .where(eq(manualRatingTable.userId, session.user.id)),
-    ],
-  );
+  const [jobResult, userStats] = await Promise.all([
+    db
+      .select({ count: count() })
+      .from(jobTable)
+      .where(eq(jobTable.userId, session.user.id)),
+    getUserStats(session.user.id),
+  ]);
 
   return {
     user: {
@@ -46,9 +33,6 @@ export const load: LayoutServerLoad = async ({ parent }) => {
     },
     isAdmin: isAdminUser(session.user.id, env.ADMIN_USER_ID),
     jobCount: jobResult[0]?.count,
-    apiKey: apiKeyResult[0]?.apiKey ?? null,
-    apiKeyCreatedAt: apiKeyResult[0]?.createdAt ?? null,
     userStats,
-    manualRatings,
   };
 };